Add files via upload
I have used the Online compiler for this i.e Snack expo. Because I attempted it in React Native and faced many issues in installing the compiler for it, So I chose Snack Expo. 
How to Run:
You have to place the File 'App.js' and images in that compiler and it will automatically Install all the dependencies by showing the message below the line. Just click the"Add dependency" and it will install.